+++ hardware/src/ara_soc.sv
@@ -8,14 +8,14 @@
 
 module ara_soc import axi_pkg::*; import ara_pkg::*; #(
     // RVV Parameters
-    parameter  int           unsigned NrLanes      = 0, // Number of parallel vector lanes.
+    parameter  int           unsigned NrLanes      = 0,                          // Number of parallel vector lanes.
     // Support for floating-point data types
     parameter  fpu_support_e          FPUSupport   = FPUSupportHalfSingleDouble,
     // AXI Interface
     parameter  int           unsigned AxiDataWidth = 32*NrLanes,
     parameter  int           unsigned AxiAddrWidth = 64,
     parameter  int           unsigned AxiUserWidth = 1,
-    parameter  int           unsigned AxiIdWidth   = 6,
+    parameter  int           unsigned AxiIdWidth   = 5,
     // Main memory
     parameter  int           unsigned L2NumWords   = 2**19,
     // Dependant parameters. DO NOT CHANGE!
@@ -43,11 +43,15 @@ module ara_soc import axi_pkg::*; import ara_pkg::*; #(
     input  logic        uart_pslverr_i
   );
 
+  `include "axi/assign.svh"
+  `include "axi/typedef.svh"
+  `include "common_cells/registers.svh"
+
   //////////////////////
   //  Memory Regions  //
   //////////////////////
 
-  localparam NrAXIMasters = 2; // Actually masters, but slaves on the crossbar
+  localparam NrAXIMasters = 1; // Actually masters, but slaves on the crossbar
 
   typedef enum int unsigned {
     L2MEM = 0,
@@ -72,9 +76,6 @@ module ara_soc import axi_pkg::*; import ara_pkg::*; #(
   //  AXI  //
   ///////////
 
-  `include "axi/assign.svh"
-  `include "axi/typedef.svh"
-
   // Ariane's AXI port data width
   localparam AxiNarrowDataWidth = 64;
   localparam AxiNarrowStrbWidth = AxiNarrowDataWidth / 8;
@@ -85,72 +86,30 @@ module ara_soc import axi_pkg::*; import ara_pkg::*; #(
   localparam AxiSocIdWidth  = AxiIdWidth - $clog2(NrAXIMasters);
   localparam AxiCoreIdWidth = AxiSocIdWidth - 1;
 
-  // AXI Typedefs
-
-  // These are used for ara_soc's AXI interface
-  `AXI_TYPEDEF_AR_CHAN_T(ar_chan_t, axi_addr_t, axi_id_t, axi_user_t)
-  `AXI_TYPEDEF_R_CHAN_T(r_chan_t, axi_data_t, axi_id_t, axi_user_t)
-  `AXI_TYPEDEF_AW_CHAN_T(aw_chan_t, axi_addr_t, axi_id_t, axi_user_t)
-  `AXI_TYPEDEF_W_CHAN_T(w_chan_t, axi_data_t, axi_strb_t, axi_user_t)
-  `AXI_TYPEDEF_B_CHAN_T(b_chan_t, axi_id_t, axi_user_t)
-  `AXI_TYPEDEF_REQ_T(axi_req_t, aw_chan_t, w_chan_t, ar_chan_t)
-  `AXI_TYPEDEF_RESP_T(axi_resp_t, b_chan_t, r_chan_t)
-
   // Internal types
   typedef logic [AxiNarrowDataWidth-1:0] axi_narrow_data_t;
   typedef logic [AxiNarrowStrbWidth-1:0] axi_narrow_strb_t;
   typedef logic [AxiSocIdWidth-1:0] axi_soc_id_t;
   typedef logic [AxiCoreIdWidth-1:0] axi_core_id_t;
 
-  `AXI_TYPEDEF_W_CHAN_T(axi_wide_w_chan_t, axi_data_t, axi_strb_t, axi_user_t)
-  `AXI_TYPEDEF_W_CHAN_T(axi_narrow_w_chan_t, axi_narrow_data_t, axi_narrow_strb_t, axi_user_t)
-
-  `AXI_TYPEDEF_AR_CHAN_T(axi_soc_ar_chan_t, axi_addr_t, axi_soc_id_t, axi_user_t)
-  `AXI_TYPEDEF_R_CHAN_T(axi_soc_wide_r_chan_t, axi_data_t, axi_soc_id_t, axi_user_t)
-  `AXI_TYPEDEF_R_CHAN_T(axi_soc_narrow_r_chan_t, axi_narrow_data_t, axi_soc_id_t, axi_user_t)
-  `AXI_TYPEDEF_AW_CHAN_T(axi_soc_aw_chan_t, axi_addr_t, axi_soc_id_t, axi_user_t)
-  `AXI_TYPEDEF_B_CHAN_T(axi_soc_b_chan_t, axi_soc_id_t, axi_user_t)
-  `AXI_TYPEDEF_REQ_T(axi_soc_wide_req_t, axi_soc_aw_chan_t, axi_wide_w_chan_t, axi_soc_ar_chan_t)
-  `AXI_TYPEDEF_RESP_T(axi_soc_wide_resp_t, axi_soc_b_chan_t, axi_soc_wide_r_chan_t)
-  `AXI_TYPEDEF_REQ_T(axi_soc_narrow_req_t, axi_soc_aw_chan_t, axi_narrow_w_chan_t,
-    axi_soc_ar_chan_t)
-  `AXI_TYPEDEF_RESP_T(axi_soc_narrow_resp_t, axi_soc_b_chan_t, axi_soc_narrow_r_chan_t)
-
-  `AXI_TYPEDEF_AR_CHAN_T(axi_core_ar_chan_t, axi_addr_t, axi_core_id_t, axi_user_t)
-  `AXI_TYPEDEF_R_CHAN_T(axi_core_wide_r_chan_t, axi_data_t, axi_core_id_t, axi_user_t)
-  `AXI_TYPEDEF_R_CHAN_T(axi_core_narrow_r_chan_t, axi_narrow_data_t, axi_core_id_t, axi_user_t)
-  `AXI_TYPEDEF_AW_CHAN_T(axi_core_aw_chan_t, axi_addr_t, axi_core_id_t, axi_user_t)
-  `AXI_TYPEDEF_B_CHAN_T(axi_core_b_chan_t, axi_core_id_t, axi_user_t)
-  `AXI_TYPEDEF_REQ_T(axi_core_wide_req_t, axi_core_aw_chan_t, axi_wide_w_chan_t, axi_core_ar_chan_t)
-  `AXI_TYPEDEF_RESP_T(axi_core_wide_resp_t, axi_core_b_chan_t, axi_core_wide_r_chan_t)
-  `AXI_TYPEDEF_REQ_T(axi_core_narrow_req_t, axi_core_aw_chan_t, axi_narrow_w_chan_t,
-    axi_core_ar_chan_t)
-  `AXI_TYPEDEF_RESP_T(axi_core_narrow_resp_t, axi_core_b_chan_t, axi_core_narrow_r_chan_t)
-
-  `AXI_LITE_TYPEDEF_AW_CHAN_T(axi_lite_soc_narrow_aw_t, axi_addr_t)
-  `AXI_LITE_TYPEDEF_W_CHAN_T(axi_lite_soc_narrow_w_t, axi_narrow_data_t, axi_narrow_strb_t)
-  `AXI_LITE_TYPEDEF_B_CHAN_T(axi_lite_soc_narrow_b_t)
-  `AXI_LITE_TYPEDEF_AR_CHAN_T(axi_lite_soc_narrow_ar_t, axi_addr_t)
-  `AXI_LITE_TYPEDEF_R_CHAN_T(axi_lite_soc_narrow_r_t, axi_narrow_data_t)
-  `AXI_LITE_TYPEDEF_REQ_T(axi_lite_soc_narrow_req_t, axi_lite_soc_narrow_aw_t,
-    axi_lite_soc_narrow_w_t, axi_lite_soc_narrow_ar_t)
-  `AXI_LITE_TYPEDEF_RESP_T(axi_lite_soc_narrow_resp_t, axi_lite_soc_narrow_b_t,
-    axi_lite_soc_narrow_r_t)
+  // AXI Typedefs
+  `AXI_TYPEDEF_ALL(system, axi_addr_t, axi_id_t, axi_data_t, axi_strb_t, axi_user_t)
+  `AXI_TYPEDEF_ALL(ara_axi, axi_addr_t, axi_core_id_t, axi_data_t, axi_strb_t, axi_user_t)
+  `AXI_TYPEDEF_ALL(ariane_axi, axi_addr_t, axi_core_id_t, axi_narrow_data_t, axi_narrow_strb_t,
+    axi_user_t)
+  `AXI_TYPEDEF_ALL(soc_narrow, axi_addr_t, axi_soc_id_t, axi_narrow_data_t, axi_narrow_strb_t,
+    axi_user_t)
+  `AXI_TYPEDEF_ALL(soc_wide, axi_addr_t, axi_soc_id_t, axi_data_t, axi_strb_t, axi_user_t)
+  `AXI_LITE_TYPEDEF_ALL(soc_narrow_lite, axi_addr_t, axi_narrow_data_t, axi_narrow_strb_t)
 
   // Buses
-  axi_core_narrow_req_t  ariane_narrow_axi_req;
-  axi_core_narrow_resp_t ariane_narrow_axi_resp;
-  axi_core_wide_req_t    ariane_axi_req;
-  axi_core_wide_resp_t   ariane_axi_resp;
-  axi_core_wide_req_t    ara_axi_req;
-  axi_core_wide_resp_t   ara_axi_resp;
-  axi_core_wide_req_t    ara_axi_req_inval;
-  axi_core_wide_resp_t   ara_axi_resp_inval;
+  system_req_t  system_axi_req;
+  system_resp_t system_axi_resp;
 
-  axi_soc_wide_req_t    [NrAXISlaves-1:0] periph_wide_axi_req;
-  axi_soc_wide_resp_t   [NrAXISlaves-1:0] periph_wide_axi_resp;
-  axi_soc_narrow_req_t  [NrAXISlaves-1:0] periph_narrow_axi_req;
-  axi_soc_narrow_resp_t [NrAXISlaves-1:0] periph_narrow_axi_resp;
+  soc_wide_req_t    [NrAXISlaves-1:0] periph_wide_axi_req;
+  soc_wide_resp_t   [NrAXISlaves-1:0] periph_wide_axi_resp;
+  soc_narrow_req_t  [NrAXISlaves-1:0] periph_narrow_axi_req;
+  soc_narrow_resp_t [NrAXISlaves-1:0] periph_narrow_axi_resp;
 
   ////////////////
   //  Crossbar  //
@@ -163,8 +122,8 @@ module ara_soc import axi_pkg::*; import ara_pkg::*; #(
     MaxSlvTrans       : 4,
     FallThrough       : 1'b0,
     LatencyMode       : axi_pkg::CUT_MST_PORTS,
-    AxiIdWidthSlvPorts: AxiCoreIdWidth,
-    AxiIdUsedSlvPorts : AxiCoreIdWidth,
+    AxiIdWidthSlvPorts: AxiSocIdWidth,
+    AxiIdUsedSlvPorts : AxiSocIdWidth,
     UniqueIds         : 1'b0,
     AxiAddrWidth      : AxiAddrWidth,
     AxiDataWidth      : AxiWideDataWidth,
@@ -180,48 +139,46 @@ module ara_soc import axi_pkg::*; import ara_pkg::*; #(
 
   axi_xbar #(
     .Cfg          (XBarCfg                ),
-    .slv_aw_chan_t(axi_core_aw_chan_t     ),
-    .mst_aw_chan_t(axi_soc_aw_chan_t      ),
-    .w_chan_t     (axi_wide_w_chan_t      ),
-    .slv_b_chan_t (axi_core_b_chan_t      ),
-    .mst_b_chan_t (axi_soc_b_chan_t       ),
-    .slv_ar_chan_t(axi_core_ar_chan_t     ),
-    .mst_ar_chan_t(axi_soc_ar_chan_t      ),
-    .slv_r_chan_t (axi_core_wide_r_chan_t ),
-    .mst_r_chan_t (axi_soc_wide_r_chan_t  ),
-    .slv_req_t    (axi_core_wide_req_t    ),
-    .slv_resp_t   (axi_core_wide_resp_t   ),
-    .mst_req_t    (axi_soc_wide_req_t     ),
-    .mst_resp_t   (axi_soc_wide_resp_t    ),
+    .slv_aw_chan_t(system_aw_chan_t       ),
+    .mst_aw_chan_t(soc_wide_aw_chan_t     ),
+    .w_chan_t     (system_w_chan_t        ),
+    .slv_b_chan_t (system_b_chan_t        ),
+    .mst_b_chan_t (soc_wide_b_chan_t      ),
+    .slv_ar_chan_t(system_ar_chan_t       ),
+    .mst_ar_chan_t(soc_wide_ar_chan_t     ),
+    .slv_r_chan_t (system_r_chan_t        ),
+    .mst_r_chan_t (soc_wide_r_chan_t      ),
+    .slv_req_t    (system_req_t           ),
+    .slv_resp_t   (system_resp_t          ),
+    .mst_req_t    (soc_wide_req_t         ),
+    .mst_resp_t   (soc_wide_resp_t        ),
     .rule_t       (axi_pkg::xbar_rule_64_t)
   ) i_soc_xbar (
-    .clk_i                (clk_i                                ),
-    .rst_ni               (rst_ni                               ),
-    .test_i               (1'b0                                 ),
-    .slv_ports_req_i      ({ariane_axi_req, ara_axi_req_inval}  ),
-    .slv_ports_resp_o     ({ariane_axi_resp, ara_axi_resp_inval}),
-    .mst_ports_req_o      (periph_wide_axi_req                  ),
-    .mst_ports_resp_i     (periph_wide_axi_resp                 ),
-    .addr_map_i           (routing_rules                        ),
-    .en_default_mst_port_i('0                                   ),
-    .default_mst_port_i   ('0                                   )
+    .clk_i                (clk_i               ),
+    .rst_ni               (rst_ni              ),
+    .test_i               (1'b0                ),
+    .slv_ports_req_i      (system_axi_req      ),
+    .slv_ports_resp_o     (system_axi_resp     ),
+    .mst_ports_req_o      (periph_wide_axi_req ),
+    .mst_ports_resp_i     (periph_wide_axi_resp),
+    .addr_map_i           (routing_rules       ),
+    .en_default_mst_port_i('0                  ),
+    .default_mst_port_i   ('0                  )
   );
 
   //////////
   //  L2  //
   //////////
 
-  `include "common_cells/registers.svh"
-
   // The L2 memory does not support atomics
 
-  axi_soc_wide_req_t  l2mem_wide_axi_req_wo_atomics;
-  axi_soc_wide_resp_t l2mem_wide_axi_resp_wo_atomics;
+  soc_wide_req_t  l2mem_wide_axi_req_wo_atomics;
+  soc_wide_resp_t l2mem_wide_axi_resp_wo_atomics;
   axi_atop_filter #(
-    .AxiIdWidth     (AxiSocIdWidth      ),
-    .AxiMaxWriteTxns(4                  ),
-    .req_t          (axi_soc_wide_req_t ),
-    .resp_t         (axi_soc_wide_resp_t)
+    .AxiIdWidth     (AxiSocIdWidth  ),
+    .AxiMaxWriteTxns(4              ),
+    .req_t          (soc_wide_req_t ),
+    .resp_t         (soc_wide_resp_t)
   ) i_l2mem_atop_filter (
     .clk_i     (clk_i                         ),
     .rst_ni    (rst_ni                        ),
@@ -240,12 +197,12 @@ module ara_soc import axi_pkg::*; import ara_pkg::*; #(
   logic                      l2_rvalid;
 
   axi_to_mem #(
-    .AddrWidth (AxiAddrWidth       ),
-    .DataWidth (AxiDataWidth       ),
-    .IdWidth   (AxiSocIdWidth      ),
-    .NumBanks  (1                  ),
-    .axi_req_t (axi_soc_wide_req_t ),
-    .axi_resp_t(axi_soc_wide_resp_t)
+    .AddrWidth (AxiAddrWidth   ),
+    .DataWidth (AxiDataWidth   ),
+    .IdWidth   (AxiSocIdWidth  ),
+    .NumBanks  (1              ),
+    .axi_req_t (soc_wide_req_t ),
+    .axi_resp_t(soc_wide_resp_t)
   ) i_axi_to_mem (
     .clk_i       (clk_i                         ),
     .rst_ni      (rst_ni                        ),
@@ -352,22 +309,22 @@ module ara_soc import axi_pkg::*; import ara_pkg::*; #(
   );
 
   axi_dw_converter #(
-    .AxiSlvPortDataWidth(AxiWideDataWidth       ),
-    .AxiMstPortDataWidth(AxiNarrowDataWidth     ),
-    .AxiAddrWidth       (AxiAddrWidth           ),
-    .AxiIdWidth         (AxiSocIdWidth          ),
-    .AxiMaxReads        (2                      ),
-    .ar_chan_t          (axi_soc_ar_chan_t      ),
-    .mst_r_chan_t       (axi_soc_narrow_r_chan_t),
-    .slv_r_chan_t       (axi_soc_wide_r_chan_t  ),
-    .aw_chan_t          (axi_soc_aw_chan_t      ),
-    .b_chan_t           (axi_soc_b_chan_t       ),
-    .mst_w_chan_t       (axi_narrow_w_chan_t    ),
-    .slv_w_chan_t       (axi_wide_w_chan_t      ),
-    .axi_mst_req_t      (axi_soc_narrow_req_t   ),
-    .axi_mst_resp_t     (axi_soc_narrow_resp_t  ),
-    .axi_slv_req_t      (axi_soc_wide_req_t     ),
-    .axi_slv_resp_t     (axi_soc_wide_resp_t    )
+    .AxiSlvPortDataWidth(AxiWideDataWidth     ),
+    .AxiMstPortDataWidth(AxiNarrowDataWidth   ),
+    .AxiAddrWidth       (AxiAddrWidth         ),
+    .AxiIdWidth         (AxiSocIdWidth        ),
+    .AxiMaxReads        (2                    ),
+    .ar_chan_t          (soc_wide_ar_chan_t   ),
+    .mst_r_chan_t       (soc_narrow_r_chan_t  ),
+    .slv_r_chan_t       (soc_wide_r_chan_t    ),
+    .aw_chan_t          (soc_narrow_aw_chan_t ),
+    .b_chan_t           (soc_wide_b_chan_t    ),
+    .mst_w_chan_t       (soc_narrow_w_chan_t  ),
+    .slv_w_chan_t       (soc_wide_w_chan_t    ),
+    .axi_mst_req_t      (soc_narrow_req_t     ),
+    .axi_mst_resp_t     (soc_narrow_resp_t    ),
+    .axi_slv_req_t      (soc_wide_req_t       ),
+    .axi_slv_resp_t     (soc_wide_resp_t      )
   ) i_axi_slave_uart_dwc (
     .clk_i     (clk_i                       ),
     .rst_ni    (rst_ni                      ),
@@ -381,21 +338,21 @@ module ara_soc import axi_pkg::*; import ara_pkg::*; #(
   //  Control registers  //
   /////////////////////////
 
-  axi_lite_soc_narrow_req_t  axi_lite_ctrl_registers_req;
-  axi_lite_soc_narrow_resp_t axi_lite_ctrl_registers_resp;
+  soc_narrow_lite_req_t  axi_lite_ctrl_registers_req;
+  soc_narrow_lite_resp_t axi_lite_ctrl_registers_resp;
 
   axi_to_axi_lite #(
-    .AxiAddrWidth   (AxiAddrWidth              ),
-    .AxiDataWidth   (AxiNarrowDataWidth        ),
-    .AxiIdWidth     (AxiSocIdWidth             ),
-    .AxiUserWidth   (AxiUserWidth              ),
-    .AxiMaxReadTxns (1                         ),
-    .AxiMaxWriteTxns(1                         ),
-    .FallThrough    (1'b0                      ),
-    .full_req_t     (axi_soc_narrow_req_t      ),
-    .full_resp_t    (axi_soc_narrow_resp_t     ),
-    .lite_req_t     (axi_lite_soc_narrow_req_t ),
-    .lite_resp_t    (axi_lite_soc_narrow_resp_t)
+    .AxiAddrWidth   (AxiAddrWidth          ),
+    .AxiDataWidth   (AxiNarrowDataWidth    ),
+    .AxiIdWidth     (AxiSocIdWidth         ),
+    .AxiUserWidth   (AxiUserWidth          ),
+    .AxiMaxReadTxns (1                     ),
+    .AxiMaxWriteTxns(1                     ),
+    .FallThrough    (1'b0                  ),
+    .full_req_t     (soc_narrow_req_t      ),
+    .full_resp_t    (soc_narrow_resp_t     ),
+    .lite_req_t     (soc_narrow_lite_req_t ),
+    .lite_resp_t    (soc_narrow_lite_resp_t)
   ) i_axi_to_axi_lite (
     .clk_i     (clk_i                        ),
     .rst_ni    (rst_ni                       ),
@@ -407,12 +364,12 @@ module ara_soc import axi_pkg::*; import ara_pkg::*; #(
   );
 
   ctrl_registers #(
-    .DRAMBaseAddr   (DRAMBase                  ),
-    .DRAMLength     (DRAMLength                ),
-    .DataWidth      (AxiNarrowDataWidth        ),
-    .AddrWidth      (AxiAddrWidth              ),
-    .axi_lite_req_t (axi_lite_soc_narrow_req_t ),
-    .axi_lite_resp_t(axi_lite_soc_narrow_resp_t)
+    .DRAMBaseAddr   (DRAMBase              ),
+    .DRAMLength     (DRAMLength            ),
+    .DataWidth      (AxiNarrowDataWidth    ),
+    .AddrWidth      (AxiAddrWidth          ),
+    .axi_lite_req_t (soc_narrow_lite_req_t ),
+    .axi_lite_resp_t(soc_narrow_lite_resp_t)
   ) i_ctrl_registers (
     .clk_i                (clk_i                       ),
     .rst_ni               (rst_ni                      ),
@@ -424,22 +381,22 @@ module ara_soc import axi_pkg::*; import ara_pkg::*; #(
   );
 
   axi_dw_converter #(
-    .AxiSlvPortDataWidth(AxiWideDataWidth       ),
-    .AxiMstPortDataWidth(AxiNarrowDataWidth     ),
-    .AxiAddrWidth       (AxiAddrWidth           ),
-    .AxiIdWidth         (AxiSocIdWidth          ),
-    .AxiMaxReads        (2                      ),
-    .ar_chan_t          (axi_soc_ar_chan_t      ),
-    .mst_r_chan_t       (axi_soc_narrow_r_chan_t),
-    .slv_r_chan_t       (axi_soc_wide_r_chan_t  ),
-    .aw_chan_t          (axi_soc_aw_chan_t      ),
-    .b_chan_t           (axi_soc_b_chan_t       ),
-    .mst_w_chan_t       (axi_narrow_w_chan_t    ),
-    .slv_w_chan_t       (axi_wide_w_chan_t      ),
-    .axi_mst_req_t      (axi_soc_narrow_req_t   ),
-    .axi_mst_resp_t     (axi_soc_narrow_resp_t  ),
-    .axi_slv_req_t      (axi_soc_wide_req_t     ),
-    .axi_slv_resp_t     (axi_soc_wide_resp_t    )
+    .AxiSlvPortDataWidth(AxiWideDataWidth    ),
+    .AxiMstPortDataWidth(AxiNarrowDataWidth  ),
+    .AxiAddrWidth       (AxiAddrWidth        ),
+    .AxiIdWidth         (AxiSocIdWidth       ),
+    .AxiMaxReads        (2                   ),
+    .ar_chan_t          (soc_wide_ar_chan_t  ),
+    .mst_r_chan_t       (soc_narrow_r_chan_t ),
+    .slv_r_chan_t       (soc_wide_r_chan_t   ),
+    .aw_chan_t          (soc_narrow_aw_chan_t),
+    .b_chan_t           (soc_narrow_b_chan_t ),
+    .mst_w_chan_t       (soc_narrow_w_chan_t ),
+    .slv_w_chan_t       (soc_wide_w_chan_t   ),
+    .axi_mst_req_t      (soc_narrow_req_t    ),
+    .axi_mst_resp_t     (soc_narrow_resp_t   ),
+    .axi_slv_req_t      (soc_wide_req_t      ),
+    .axi_slv_resp_t     (soc_wide_resp_t     )
   ) i_axi_slave_ctrl_dwc (
     .clk_i     (clk_i                       ),
     .rst_ni    (rst_ni                      ),
@@ -449,12 +406,9 @@ module ara_soc import axi_pkg::*; import ara_pkg::*; #(
     .mst_resp_i(periph_narrow_axi_resp[CTRL])
   );
 
-  //////////////////////
-  //  Ara and Ariane  //
-  //////////////////////
-
-  import ariane_pkg::accelerator_req_t;
-  import ariane_pkg::accelerator_resp_t;
+  //////////////
+  //  System  //
+  //////////////
 
   localparam ariane_pkg::ariane_cfg_t ArianeAraConfig = '{
     RASDepth             : 2,
@@ -480,116 +434,44 @@ module ara_soc import axi_pkg::*; import ara_pkg::*; #(
     NrPMPEntries         : 0
   };
 
-  // Accelerator ports
-  accelerator_req_t                     acc_req;
-  logic                                 acc_req_valid;
-  logic                                 acc_req_ready;
-  accelerator_resp_t                    acc_resp;
-  logic                                 acc_resp_valid;
-  logic                                 acc_resp_ready;
-  logic                                 acc_cons_en;
-  logic              [AxiAddrWidth-1:0] inval_addr;
-  logic                                 inval_valid;
-  logic                                 inval_ready;
-
-  ariane #(
-    .ArianeCfg(ArianeAraConfig)
-  ) i_ariane (
-    .clk_i            (clk_i                 ),
-    .rst_ni           (rst_ni                ),
-    .boot_addr_i      (DRAMBase              ), // start fetching from DRAM
-    .hart_id_i        ('0                    ),
-    .irq_i            ('0                    ),
-    .ipi_i            ('0                    ),
-    .time_irq_i       ('0                    ),
-    .debug_req_i      ('0                    ),
-    .axi_req_o        (ariane_narrow_axi_req ),
-    .axi_resp_i       (ariane_narrow_axi_resp),
-    // Accelerator ports
-    .acc_req_o        (acc_req               ),
-    .acc_req_valid_o  (acc_req_valid         ),
-    .acc_req_ready_i  (acc_req_ready         ),
-    .acc_resp_i       (acc_resp              ),
-    .acc_resp_valid_i (acc_resp_valid        ),
-    .acc_resp_ready_o (acc_resp_ready        ),
-    .acc_cons_en_o    (acc_cons_en           ),
-    .inval_addr_i     (inval_addr            ),
-    .inval_valid_i    (inval_valid           ),
-    .inval_ready_o    (inval_ready           )
-  );
-
-  axi_dw_converter #(
-    .AxiSlvPortDataWidth(AxiNarrowDataWidth      ),
-    .AxiMstPortDataWidth(AxiWideDataWidth        ),
-    .AxiAddrWidth       (AxiAddrWidth            ),
-    .AxiIdWidth         (AxiCoreIdWidth          ),
-    .AxiMaxReads        (4                       ),
-    .ar_chan_t          (axi_core_ar_chan_t      ),
-    .mst_r_chan_t       (axi_core_wide_r_chan_t  ),
-    .slv_r_chan_t       (axi_core_narrow_r_chan_t),
-    .aw_chan_t          (axi_core_aw_chan_t      ),
-    .b_chan_t           (axi_core_b_chan_t       ),
-    .mst_w_chan_t       (axi_wide_w_chan_t       ),
-    .slv_w_chan_t       (axi_narrow_w_chan_t     ),
-    .axi_mst_req_t      (axi_core_wide_req_t     ),
-    .axi_mst_resp_t     (axi_core_wide_resp_t    ),
-    .axi_slv_req_t      (axi_core_narrow_req_t   ),
-    .axi_slv_resp_t     (axi_core_narrow_resp_t  )
-  ) i_ariane_axi_dwc (
-    .clk_i     (clk_i                 ),
-    .rst_ni    (rst_ni                ),
-    .slv_req_i (ariane_narrow_axi_req ),
-    .slv_resp_o(ariane_narrow_axi_resp),
-    .mst_req_o (ariane_axi_req        ),
-    .mst_resp_i(ariane_axi_resp       )
-  );
-
-  axi_inval_filter #(
-    .MaxTxns    (4                              ),
-    .AddrWidth  (AxiAddrWidth                   ),
-    .L1LineWidth(ariane_pkg::DCACHE_LINE_WIDTH/8),
-    .aw_chan_t  (axi_core_aw_chan_t             ),
-    .req_t      (axi_core_wide_req_t            ),
-    .resp_t     (axi_core_wide_resp_t           )
-  ) i_axi_inval_filter (
-    .clk_i        (clk_i             ),
-    .rst_ni       (rst_ni            ),
-    .en_i         (acc_cons_en       ),
-    .slv_req_i    (ara_axi_req       ),
-    .slv_resp_o   (ara_axi_resp      ),
-    .mst_req_o    (ara_axi_req_inval ),
-    .mst_resp_i   (ara_axi_resp_inval),
-    .inval_addr_o (inval_addr        ),
-    .inval_valid_o(inval_valid       ),
-    .inval_ready_i(inval_ready       )
-  );
-
-  ara #(
-    .NrLanes     (NrLanes               ),
-    .FPUSupport  (FPUSupport            ),
-    .AxiDataWidth(AxiWideDataWidth      ),
-    .AxiAddrWidth(AxiAddrWidth          ),
-    .axi_ar_t    (axi_core_ar_chan_t    ),
-    .axi_r_t     (axi_core_wide_r_chan_t),
-    .axi_aw_t    (axi_core_aw_chan_t    ),
-    .axi_w_t     (axi_wide_w_chan_t     ),
-    .axi_b_t     (axi_core_b_chan_t     ),
-    .axi_req_t   (axi_core_wide_req_t   ),
-    .axi_resp_t  (axi_core_wide_resp_t  )
-  ) i_ara (
-    .clk_i           (clk_i         ),
-    .rst_ni          (rst_ni        ),
-    .scan_enable_i   (scan_enable_i ),
-    .scan_data_i     (1'b0          ),
-    .scan_data_o     (/* Unused */  ),
-    .acc_req_i       (acc_req       ),
-    .acc_req_valid_i (acc_req_valid ),
-    .acc_req_ready_o (acc_req_ready ),
-    .acc_resp_o      (acc_resp      ),
-    .acc_resp_valid_o(acc_resp_valid),
-    .acc_resp_ready_i(acc_resp_ready),
-    .axi_req_o       (ara_axi_req   ),
-    .axi_resp_i      (ara_axi_resp  )
+  ara_system #(
+    .NrLanes           (NrLanes              ),
+    .FPUSupport        (FPUSupport           ),
+    .ArianeCfg         (ArianeAraConfig      ),
+    .AxiAddrWidth      (AxiAddrWidth         ),
+    .AxiIdWidth        (AxiCoreIdWidth       ),
+    .AxiNarrowDataWidth(AxiNarrowDataWidth   ),
+    .AxiWideDataWidth  (AxiDataWidth         ),
+    .ara_axi_ar_t      (ara_axi_ar_chan_t    ),
+    .ara_axi_aw_t      (ara_axi_aw_chan_t    ),
+    .ara_axi_b_t       (ara_axi_b_chan_t     ),
+    .ara_axi_r_t       (ara_axi_r_chan_t     ),
+    .ara_axi_w_t       (ara_axi_w_chan_t     ),
+    .ara_axi_req_t     (ara_axi_req_t        ),
+    .ara_axi_resp_t    (ara_axi_resp_t       ),
+    .ariane_axi_ar_t   (ariane_axi_ar_chan_t ),
+    .ariane_axi_aw_t   (ariane_axi_aw_chan_t ),
+    .ariane_axi_b_t    (ariane_axi_b_chan_t  ),
+    .ariane_axi_r_t    (ariane_axi_r_chan_t  ),
+    .ariane_axi_w_t    (ariane_axi_w_chan_t  ),
+    .ariane_axi_req_t  (ariane_axi_req_t     ),
+    .ariane_axi_resp_t (ariane_axi_resp_t    ),
+    .system_axi_ar_t   (system_ar_chan_t     ),
+    .system_axi_aw_t   (system_aw_chan_t     ),
+    .system_axi_b_t    (system_b_chan_t      ),
+    .system_axi_r_t    (system_r_chan_t      ),
+    .system_axi_w_t    (system_w_chan_t      ),
+    .system_axi_req_t  (system_req_t         ),
+    .system_axi_resp_t (system_resp_t        )
+  ) i_system (
+    .clk_i        (clk_i            ),
+    .rst_ni       (rst_ni           ),
+    .boot_addr_i  (DRAMBase         ), // start fetching from DRAM
+    .scan_enable_i(1'b0             ),
+    .scan_data_i  (1'b0             ),
+    .scan_data_o  (/* Unconnected */),
+    .axi_req_o    (system_axi_req   ),
+    .axi_resp_i   (system_axi_resp  )
   );
 
   //////////////////